WIP MDEV-17380 innodb_flush_neighbors=ON should be ignored on SSD
For tablespaces that do not reside on spinning storage, it does
not make sense to attempt to write nearby pages when writing out
dirty pages from the InnoDB buffer pool. It is actually detrimental
to performance and to the life span of flash ROM storage.

This patch implements the SSD detection on Windows, but it does not
do that on other systems yet.

TODO: On Linux, at startup, construct a mapping of st_dev that reside
on non-rotational media, by joining /proc/self/mountinfo with the output of
grep -l '^0$'/sys/block/*/queue/rotational

Then, in fil_node_t::read_page0(), if the statbuf.st_dev is found in
the mapping, we would set the fil_node_t::on_ssd flag.